Replace all literal occurrences of actual* with current*
I've replaced all occurrences of "actual" and variants like "actually"
in string literals with "current" and "currently".  This usage of
"actual" is a common mistake because it looks like the German word
"aktuell", but for information that is fresh and up-to-date about the
present we use "current", whereas "actually" is closer to "eigentlich".